+++ CHANGELOG.md
@@ -8,6 +8,11 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Adds
+
+- Adds the `"connect"` section hues to the light and dark color palettes.
+- Adds the `"connect"` color variant to the `Breadcrumbs` component.
+
 ## 1.6.0 (June 8, 2023)
 
 ### Vite Foundation Update